refactor: add type file
