+++ src/components/config/SelectQuizSet.tsx
@@ -15,7 +15,8 @@ import {
 } from "@chakra-ui/react";
 import { Upload } from "tabler-icons-react";
 
-import db, { GameDBQuizProps } from "#/utils/db";
+import db from "#/utils/db";
+import { GameDBQuizProps } from "#/utils/types";
 
 type SelectQuizsetProps = {
   game_id: string;
