+++ src/components/home/OtherRules.tsx
@@ -2,18 +2,6 @@ import { Link as ReactLink } from "react-router-dom";
 
 import { Box, Link, ListItem, UnorderedList } from "@chakra-ui/react";
 
-export type AQLGameProps = {
-  id: string;
-  name: string;
-  left_team: string;
-  right_team: string;
-  quiz: {
-    set_name: string;
-    offset: number;
-  };
-  last_open: string;
-};
-
 const OtherRules: React.FC = () => {
   return (
     <Box pt={5}>
