+++ src/utils/db.ts
@@ -1,107 +1,6 @@
-import Dexie, { Table } from "dexie";
+import Dexie from "dexie";
 
-export type DexieDatabase = { [P in keyof Dexie]: Dexie[P] };
-
-export type RuleNames =
-  | "normal"
-  | "nomx"
-  | "nomx-ad"
-  | "ny"
-  | "nomr"
-  | "nbyn"
-  | "nupdown"
-  | "swedish10"
-  | "backstream"
-  | "attacksurvival"
-  | "squarex"
-  | "z"
-  | "freezex"
-  | "various-fluctuations";
-
-export type GameDBPlayerProps = {
-  id: string;
-  name: string;
-  initial_correct: number;
-  initial_wrong: number;
-  base_correct_point: number;
-  base_wrong_point: number;
-};
-
-export type GameDBQuizProps = {
-  set_name: string;
-  offset: number;
-};
-
-export type GameDBProps = {
-  id: string;
-  name: string;
-  players: GameDBPlayerProps[];
-  rule: RuleNames;
-  correct_me: number;
-  wrong_me: number;
-  correct_other?: number;
-  wrong_other?: number;
-  win_point?: number;
-  lose_point?: number;
-  win_through?: number;
-  limit?: number;
-  quiz?: GameDBQuizProps;
-  discord_webhook_url?: string;
-  editable: boolean;
-  last_open: string;
-};
-
-export type PlayerDBProps = {
-  id: string;
-  name: string;
-  text: string;
-  belong: string;
-  tags: string[];
-};
-
-export type Variants = "correct" | "wrong" | "through";
-export type LogDBProps = {
-  id: string;
-  game_id: string;
-  player_id: string;
-  variant: Variants;
-  system: boolean;
-  timestamp: string;
-};
-
-export type States = "win" | "lose" | "playing";
-export type ComputedScoreDBProps = {
-  game_id: string;
-  player_id: string;
-  state: States;
-  reachState: States;
-  score: number;
-  correct: number; // 正解数
-  wrong: number; // 誤答数
-  last_correct: number; // 最後に正解した問題番号
-  last_wrong: number; // 最後に誤答した問題番号
-  odd_score: number; // 奇数問目のスコア
-  even_score: number; // 偶数問目のスコア
-  stage: number;
-  isIncapacity: boolean;
-  order: number; // プレイヤー同士の評価順
-  text: string; // 画面上に表示するための文字
-};
-
-export type QuizDBProps = {
-  id: string;
-  n: number;
-  q: string;
-  a: string;
-  set_name: string;
-};
-
-export interface ScoreWatcherDBTables extends DexieDatabase {
-  games: Table<GameDBProps>;
-  players: Table<PlayerDBProps>;
-  logs: Table<LogDBProps>;
-  quizes: Table<QuizDBProps>;
-}
+import { ScoreWatcherDBTables } from "./types";
 
 const db = new Dexie("score_watcher") as ScoreWatcherDBTables;
 db.version(2).stores({
